# docs/swagger.yaml
basePath: /
definitions:
  ErrorResponse:
    properties:
      message:
        type: string
      param:
        type: string
      value:
        type: string
    type: object
  controllers.SquadResponse:
    type: object
...
  models.Squad:
    properties:
      city:
        type: string
      name:
        type: string
      squadId:
        type: integer
    type: object
  models.SquadPlayerDetails:
    properties:
      forename:
        type: string
      nickname:
        type: string
      playerId:
        type: integer
      surname:
        type: string
    type: object
...
  payloads.Login:
    properties:
      email:
        maxLength: 100
        minLength: 5
        type: string
      password:
        maxLength: 25
        minLength: 5
        type: string
    required:
    - email
    - password
    type: object
...
  payloads.Refresh:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  payloads.Register:
    properties:
      email:
        maxLength: 100
        minLength: 5
        type: string
      forename:
        maxLength: 50
        minLength: 3
        type: string
      password:
        maxLength: 25
        minLength: 5
        type: string
      surname:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - forename
    - password
    - surname
    type: object
  payloads.RegisterResponse:
    properties:
      id:
        type: integer
    type: object
  payloads.SquadRequest:
    properties:
      city:
        maxLength: 50
        minLength: 3
        type: string
      name:
        maxLength: 30
        minLength: 2
        type: string
    required:
    - city
    - name
    type: object
  payloads.TokenPairResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
...
host: 127.0.0.1:3000
info:
  contact:
    email: nathan.mcalinden@unosquare.com
    name: Footpal Support
  description: This is swagger for Footpal
  termsOfService: http://swagger.io/terms/
  title: Footpal API
  version: "1.0"
paths:
...
  /deactivate:
    delete:
      description: Delete user from Footpal
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - user
  /login:
    post:
      description: Login to Footpal
      parameters:
      - description: Request
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/payloads.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payloads.TokenPairResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      tags:
      - user
...
  /refresh:
    post:
      description: Refresh token
      parameters:
      - description: Request
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/payloads.Refresh'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payloads.TokenPairResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      tags:
      - user
  /register:
    post:
      description: Register as new player
      parameters:
      - description: Request
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/payloads.Register'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payloads.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      tags:
      - user
  /squads:
    get:
      description: Retrieve all squads
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Squad'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      tags:
      - squad
    post:
      description: Create new squad for building up a team of players
      parameters:
      - description: Request
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/payloads.SquadRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controllers.SquadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - squad
  /squads/{squadId}:
    delete:
      description: Delete squad from footpal
      parameters:
      - description: Squad ID
        in: path
        name: squadId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: accepted
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - squad
    get:
      description: Get squad by squadId
      parameters:
      - description: Squad ID
        in: path
        name: squadId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Squad'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      tags:
      - squad
    put:
      description: Update Squad Info
      parameters:
      - description: Request
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/payloads.SquadRequest'
      - description: Squad ID
        in: path
        name: squadId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Squad'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - squad
  /squads/{squadId}/players:
    get:
      description: List of players associated with squad
      parameters:
      - description: Squad ID
        in: path
        name: squadId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SquadPlayerDetails'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      tags:
      - squad
  /squads/{squadId}/players/{playerId}:
    delete:
      description: Remove player from squad
      parameters:
      - description: Squad ID
        in: path
        name: squadId
        required: true
        type: integer
      - description: Player ID
        in: path
        name: playerId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - squad
    put:
      description: Approve player to join squad
      parameters:
      - description: Squad ID
        in: path
        name: squadId
        required: true
        type: integer
      - description: Player ID
        in: path
        name: playerId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - squad
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"